# src/main/resources/open-api/api.yaml
openapi: 3.0.3
info:
  title: MP-Monitoring API
  description: |-
    Monitoring System API for storing and retrieving time-series metrics with metadata.
  contact:
    email: udalov216@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: localhost:8082
    description: localhost.
paths:
...
  /project:
    post:
      summary: Create a new project
      operationId: createProject
      tags:
        - project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
      responses:
        '200':
          description: Successfully created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /project/{projectName}:
    get:
      summary: Get project by name
      operationId: getProject
      tags:
        - project
      parameters:
        - name: projectName
          in: path
          required: true
          description: The name of the project.
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  ## METRICS ENDPOINT
...
components:
  schemas:
...
    ProjectDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the project.
        description:
          type: string
          description: A description of the project.
      required:
        - name
        - description
...
    ApiErrorResponse:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          readOnly: true
        code:
          type: string
          readOnly: true
        exceptionName:
          type: string
          readOnly: true
        exceptionMessage:
          type: string
          readOnly: true
        stacktrace:
          type: array
          items:
            type: string